refactor: altera nome do arquivo UserContext e atualiza importações
