+++ apps/frontend/src/app/(pages)/layout.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { UserProvider } from '@/data/contexts/UseContext'
+import { UserProvider } from '@/data/contexts/UserContext'
 
 export default function Layout({ children }: any) {
     return <UserProvider>{children}</UserProvider>
