+++ apps/frontend/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import { UserProvider } from "@/data/contexts/UseContext";
+import { UserProvider } from "@/data/contexts/UserContext";
 
 const inter = Inter({ subsets: ["latin"] });
