+++ apps/frontend/src/data/hooks/useUser.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react'
-import UserContext from '../contexts/UseContext'
+import UserContext from '../contexts/UserContext'
 
 const useUser = () => useContext(UserContext)
 export default useUser
